# openapi.yaml
openapi: 3.0.3

info:
  title: Readify Bookstore API
  version: 1.0.0
  description: RESTful API for managing books and authors.

servers:
  - url: http://127.0.0.1:5000

paths:

  /books:
    get:
      summary: List all books
      responses:
        "200":
          description: Books retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"

    post:
      summary: Create a new book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookInput"

      responses:
        "201":
          description: Book created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"

        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /books/{id}:

    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer

    get:
      summary: Retrieve a book by ID
      responses:
        "200":
          description: Book retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"

        "404":
          description: Book not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      summary: Replace a book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookInput"

      responses:
        "200":
          description: Book updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"

        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        "404":
          description: Book not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      summary: Partially update a book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookPatch"

      responses:
        "200":
          description: Book updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"

        "404":
          description: Book not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete a book
      responses:
        "200":
          description: Book deleted successfully

        "404":
          description: Book not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /authors:
    get:
      summary: List all authors
      responses:
        "200":
          description: Authors retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Author"

  /health:
    get:
      summary: Check API health
      responses:
        "200":
          description: API is healthy

components:

  schemas:

    Book:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: The Great Gatsby
        author:
          type: string
          example: F. Scott Fitzgerald

    BookInput:
      type: object
      required:
        - title
        - author
      properties:
        title:
          type: string
          example: The Great Gatsby
        author:
          type: string
          example: F. Scott Fitzgerald

    BookPatch:
      type: object
      properties:
        title:
          type: string
        author:
          type: string

    Author:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: George Orwell

    Error:
      type: object
      properties:
        error:
          type: string
          example: Book not found